+++ src/py/flwr/common/constant.py
@@ -37,6 +37,8 @@ TRANSPORT_TYPES = [
     TRANSPORT_TYPE_VCE,
 ]
 
+GRPC_MAX_MESSAGE_LENGTH: int = 536_870_912  # == 512 * 1024 * 1024
+
 # Addresses
 # Ports
 CLIENTAPPIO_PORT = "9094"
